fix: move command and event list and delete comments
